feat: ui fixes

+++ src/components/sidepanel/PopulatedState.tsx
@@ -446,15 +446,15 @@ export function PopulatedState({ onShowEmpty, initialTab }: PopulatedStateProps)
         let label = ''
         if (sessionDay.getTime() === today.getTime()) {
           const dayName = sessionDay.toLocaleString('en-US', { weekday: 'long' })
-          const dateStr = sessionDay.toLocaleString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })
+          const dateStr = sessionDay.toLocaleString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
           label = `Today - ${dayName}, ${dateStr}`
         } else if (sessionDay.getTime() === yesterday.getTime()) {
           const dayName = sessionDay.toLocaleString('en-US', { weekday: 'long' })
-          const dateStr = sessionDay.toLocaleString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })
+          const dateStr = sessionDay.toLocaleString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
           label = `Yesterday - ${dayName}, ${dateStr}`
         } else {
           const dayName = sessionDay.toLocaleString('en-US', { weekday: 'long' })
-          const dateStr = sessionDay.toLocaleString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })
+          const dateStr = sessionDay.toLocaleString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
           label = `${dayName}, ${dateStr}`
         }
 
@@ -519,15 +519,15 @@ export function PopulatedState({ onShowEmpty, initialTab }: PopulatedStateProps)
         let label = ''
         if (clusterDay.getTime() === today.getTime()) {
           const dayName = clusterDay.toLocaleString('en-US', { weekday: 'long' })
-          const dateStr = clusterDay.toLocaleString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })
+          const dateStr = clusterDay.toLocaleString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
           label = `Today - ${dayName}, ${dateStr}`
         } else if (clusterDay.getTime() === yesterday.getTime()) {
           const dayName = clusterDay.toLocaleString('en-US', { weekday: 'long' })
-          const dateStr = clusterDay.toLocaleString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })
+          const dateStr = clusterDay.toLocaleString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
           label = `Yesterday - ${dayName}, ${dateStr}`
         } else {
           const dayName = clusterDay.toLocaleString('en-US', { weekday: 'long' })
-          const dateStr = clusterDay.toLocaleString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })
+          const dateStr = clusterDay.toLocaleString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
           label = `${dayName}, ${dateStr}`
         }
